requirements.txt
asgiref==3.8.1
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==3.0.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.1
requests==2.31.0
sqlparse==0.5.3
tzdata==2025.1
celery>=5.3.0
redis>=4.5.0

# Supabase y dependencias relacionadas
supabase==2.3.1
gotrue==2.1.0
supafunc==0.3.1
postgrest==0.13.0
realtime==1.0.0
storage3==0.7.0

# Utilidades
tqdm==4.66.2
python-magic==0.4.27
python-magic-bin==0.4.14; platform_system=='Windows'
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# Dependencias para Vercel
uvicorn==0.27.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
# Libreria para estructura de proyecto
repomix==0.2.2

# openai-agents==0.0.11  # Ya no se usa, reemplazado por la API directa de OpenAI
openai==1.75.0
